re-organise commands, replace magic strings with ExperimentArtefacts class
