+++ deet/scripts/cli.py
@@ -2,20 +2,11 @@
 """A CLI app to run deet pipelines."""
 
 import warnings
-from pathlib import Path
-from typing import Annotated
 
 import typer
 
-from deet.data_models.enums import CustomPromptPopulationMethod
-from deet.data_models.project import DeetProject
 from deet.logger import logger
-from deet.processors.converter_register import SupportedImportFormat
-from deet.scripts.commands import project
-from deet.settings import LogLevel
-from deet.ui import notify
-from deet.ui.wizards import run_model_wizard
-from deet.utils.cli_utils import fail_with_message
+from deet.scripts.commands import project, run
 
 APP_HELP = (
     "deet (data extraction evaluation toolkit) 🚤\n\n"
@@ -33,243 +24,7 @@ APP_HELP = (
 app = typer.Typer(help=APP_HELP, add_completion=True)
 
 app.add_typer(project.app, name="project")
-
-# Shared argument definitions and defaults
-
-GS_DATA_PATH = Annotated[
-    Path,
-    typer.Argument(..., help="Path to gold standard annotation file."),
-]
-
-DEFAULT_IMPORT_FORMAT = SupportedImportFormat.EPPI_JSON
-
-GS_DATA_FORMAT = Annotated[
-    SupportedImportFormat,
-    typer.Option(
-        help="Format of the input data (determines which converter to use)",
-    ),
-]
-
-
-LINK_MAP_PATH = Annotated[
-    Path,
-    typer.Option(
-        help="Path to write the link map",
-    ),
-]
-
-LINK_MAP_PATH_READ = Annotated[
-    Path,
-    typer.Option(
-        help="A path to a link map (create this by running "
-        "`deet init-linkage-mapping-file`)"
-    ),
-]
-
-DEFAULT_METRICS_CSV = Path("metrics.csv")
-DEFAULT_OUTPUT_COMPARISON_CSV = Path("goldstandard_llm_comparison.csv")
-
-
-@app.command()
-def link_documents_fulltexts() -> None:
-    """
-    Link documents to their fulltexts.
-
-    This creates a document containing the parsed output of its corresponding
-    fulltext in the folder defined in `output_path`. Linking will be
-    attempted using a mapping file, if provided, then by matching the
-    filename with author and year, then by matching by document id. See
-    `deet.processors.linker` for more details.
-
-    """
-    from deet.processors.linker import DocumentReferenceLinker, LinkingStrategy
-
-    deet_project = DeetProject.load()
-    processed_annotation_data = deet_project.process_data()
-
-    linker = DocumentReferenceLinker(
-        references=processed_annotation_data.documents,
-        document_base_dir=deet_project.pdf_dir,
-        document_reference_mapping=deet_project.link_map_path,
-        linking_strategies=[LinkingStrategy.MAPPING_FILE],
-    )
-    linked_documents = linker.link_many_references_parsed_documents()
-
-    if not deet_project.linked_documents_path.exists():
-        deet_project.linked_documents_path.mkdir()
-
-    if len(linked_documents) == 0:
-        fail_with_message("Error. Could not link any documents!")
-
-    for linked_document in linked_documents:
-        file_path = (
-            deet_project.linked_documents_path
-            / f"{linked_document.safe_identity.document_id}.json"
-        )
-        linked_document.save(file_path)
-
-
-@app.command()
-def extract_data(
-    config_path: Annotated[
-        Path | None,
-        typer.Option(
-            help="A path to a config file containing options for data "
-            "extraction config. A template can be generated by running "
-            "`deet export-config-template."
-        ),
-    ] = None,
-    prompt_population: Annotated[
-        CustomPromptPopulationMethod | None,
-        typer.Option(
-            help="A method to define custom prompts for your attributes to be "
-            "extracted. Leave blank to use the prompts in your gold standard "
-            "data. Set to `file` to provide a file of prompt definitions "
-            "(make sure this is supplied below). Set to `cli` to define prompts"
-            " interactively in the CLI. With `file`, only attributes that appear "
-            "in the CSV with a non-empty `prompt` are kept for extraction and "
-            "evaluation (see also `--csv-path`)."
-        ),
-    ] = CustomPromptPopulationMethod.FILE,
-    run_name: Annotated[
-        str,
-        typer.Option(
-            help="A name for the run (which will appended to a timestamp) "
-            "to help you identify this run later"
-        ),
-    ] = "",
-    custom_evaluation_metrics: Annotated[
-        list[str] | None,
-        typer.Option(
-            help="A list of additional sklearn metrics that you wish to "
-            " calculate. Use this option for each additional metric you "
-            " would like to add, e.g. `deet extract-data "
-            "--custom-evaluation-metrics brier_score_loss "
-            "--custom-evaluation-metrics cohen_kappa_score`"
-        ),
-    ] = None,
-) -> None:
-    """
-    Extract data from documents and evaluate.
-
-    Load gold standard annotation data, and use an LLM to extract data from the
-    documents in your dataset. Evaluate by comparing the results to the gold
-    standard data.
-    """
-    import yaml
-
-    from deet.evaluators.gold_standard_llm_evaluator import GoldStandardLLMEvaluator
-    from deet.extractors.cli_helpers import (
-        init_extraction_run,
-        load_or_init_config,
-        prepare_documents,
-    )
-    from deet.extractors.llm_data_extractor import (
-        DataExtractionConfig,
-        LLMDataExtractor,
-    )
-
-    if config_path is None:
-        config = run_model_wizard(DataExtractionConfig)
-    else:
-        config = load_or_init_config(config_path=config_path)
-
-    deet_project = DeetProject.load()
-    processed_annotation_data = deet_project.process_data()
-
-    extraction_run_id, experiment_out_dir = init_extraction_run(
-        deet_project.out_dir, run_name
-    )
-
-    if prompt_population is not None:
-        processed_annotation_data.populate_custom_prompts(
-            method=prompt_population, filepath=deet_project.prompt_csv_path
-        )
-
-    data_extractor = LLMDataExtractor(config=config)
-
-    documents = prepare_documents(
-        processed_annotation_data.documents,
-        config,
-        linked_document_path=deet_project.linked_documents_path,
-        pdf_dir=deet_project.pdf_dir,
-        link_map_path=deet_project.link_map_path,
-    )
-
-    run_output = data_extractor.extract_from_documents(
-        attributes=processed_annotation_data.attributes,
-        documents=documents,
-        context_type=data_extractor.config.default_context_type,
-        output_file=experiment_out_dir / "annotated_docs.json",
-        show_progress=True,
-    )
-
-    prompts_out = experiment_out_dir / "prompts.csv"
-    processed_annotation_data.export_attributes_csv_file(prompts_out)
-
-    config_out = experiment_out_dir / "config.yaml"
-    config_out.write_text(
-        yaml.safe_dump(data_extractor.config.model_dump(mode="json"), sort_keys=False),
-        encoding="utf-8",
-    )
-
-    evaluator = GoldStandardLLMEvaluator(
-        gold_standard_annotated_documents=processed_annotation_data.annotated_documents,
-        llm_annotated_documents=run_output.annotated_documents,
-        attributes=processed_annotation_data.attributes,
-        custom_metrics=custom_evaluation_metrics,
-        extraction_run_id=extraction_run_id,
-    )
-    evaluator.evaluate_llm_annotations()
-    evaluator.write_metrics_to_csv(experiment_out_dir / DEFAULT_METRICS_CSV)
-    evaluator.export_llm_comparison(experiment_out_dir / DEFAULT_OUTPUT_COMPARISON_CSV)
-    evaluator.display_metrics()
-
-
-@app.command()
-def test_llm_config(
-    config_path: Annotated[
-        Path,
-        typer.Option(
-            help="A path to a config file containing options for data "
-            "extraction config. A template can be generated by running "
-            "`deet export-config-template."
-        ),
-    ] = project.DEFAULT_CONFIG_PATH,
-) -> None:
-    """Test llm config."""
-    from deet.data_models.base import Attribute, AttributeType
-    from deet.extractors.cli_helpers import (
-        load_or_init_config,
-    )
-    from deet.extractors.llm_data_extractor import (
-        LLMDataExtractor,
-    )
-
-    config = load_or_init_config(config_path=config_path)
-    data_extractor = LLMDataExtractor(config=config)
-    attr = Attribute(
-        output_data_type=AttributeType.BOOL,
-        attribute_id=1234,
-        attribute_label="Test Attribute",
-        prompt="Is the document about climate and health? Return a BOOL",
-    )
-    context = (
-        "This is document, extract data from me please. I am about climate and health"
-    )
-    response = data_extractor.extract_from_document(
-        attributes=[attr],
-        payload=context,
-        context_type=None,
-    )
-    if response.annotations:
-        notify(
-            (
-                f"Successfully returned {len(response.annotations)} annotation: "
-                f"{response.annotations}"
-            ),
-            level=LogLevel.SUCCESS,
-        )
+app.add_typer(run.app, name="run")
 
 
 @app.callback()
