+++ deet/scripts/commands/project.py
@@ -1,6 +1,7 @@
-"""Sub-commands for project initialisation wizard."""
+"""Sub-commands for project initialisation and configuration."""
 
 from pathlib import Path
+from typing import Annotated
 
 import typer
 from InquirerPy import inquirer
@@ -86,3 +87,88 @@ def init() -> None:
     project.populate_env()
 
     export_config_template()
+
+
+@app.command()
+def link_documents_fulltexts() -> None:
+    """
+    Link documents to their fulltexts.
+
+    This creates a document containing the parsed output of its corresponding
+    fulltext in the folder defined in `output_path`. Linking will be
+    attempted using a mapping file, if provided, then by matching the
+    filename with author and year, then by matching by document id. See
+    `deet.processors.linker` for more details.
+
+    """
+    from deet.processors.linker import DocumentReferenceLinker, LinkingStrategy
+
+    deet_project = DeetProject.load()
+    processed_annotation_data = deet_project.process_data()
+
+    linker = DocumentReferenceLinker(
+        references=processed_annotation_data.documents,
+        document_base_dir=deet_project.pdf_dir,
+        document_reference_mapping=deet_project.link_map_path,
+        linking_strategies=[LinkingStrategy.MAPPING_FILE],
+    )
+    linked_documents = linker.link_many_references_parsed_documents()
+
+    if not deet_project.linked_documents_path.exists():
+        deet_project.linked_documents_path.mkdir()
+
+    if len(linked_documents) == 0:
+        fail_with_message("Error. Could not link any documents!")
+
+    for linked_document in linked_documents:
+        file_path = (
+            deet_project.linked_documents_path
+            / f"{linked_document.safe_identity.document_id}.json"
+        )
+        linked_document.save(file_path)
+
+
+@app.command()
+def test_llm_config(
+    config_path: Annotated[
+        Path,
+        typer.Option(
+            help="A path to a config file containing options for data "
+            "extraction config. A template can be generated by running "
+            "`deet export-config-template."
+        ),
+    ] = DEFAULT_CONFIG_PATH,
+) -> None:
+    """Test llm config."""
+    from deet.data_models.base import Attribute, AttributeType
+    from deet.extractors.cli_helpers import (
+        load_or_init_config,
+    )
+    from deet.extractors.llm_data_extractor import (
+        LLMDataExtractor,
+    )
+
+    config = load_or_init_config(config_path=config_path)
+    data_extractor = LLMDataExtractor(config=config)
+    attr = Attribute(
+        output_data_type=AttributeType.BOOL,
+        attribute_id=1234,
+        attribute_label="Test Attribute",
+        prompt="Is the document about climate and health? Return a BOOL",
+    )
+    context = (
+        "This is document, extract data from me please. I am about climate and health"
+    )
+    response = data_extractor.extract_from_document(
+        attributes=[attr],
+        payload=context,
+        context_type=None,
+    )
+    if response.annotations:
+        notify(
+            (
+                f"Successfully returned {len(response.annotations)} annotation: "
+                f"{response.annotations}"
+            ),
+            level=LogLevel.SUCCESS,
+        )
